refactor exploration editor/player test suites
